+++ flutter_extensions/lib/src/regexp.dart
@@ -1,4 +1,4 @@
-extension RegExpMatchExt on RegExpMatch {
+extension RegExpFlutterExtensions on RegExpMatch {
   String? namedGroupOrNull(String name) {
     try {
       return namedGroup(name);