+++ flutter_extensions/lib/src/stream.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:rxdart/rxdart.dart';
 
-extension StreamFlutterExt<T> on Stream<T> {
+extension StreamFlutterExtensions<T> on Stream<T> {
   Stream<T> dumpErrorToConsoleFlutter() => doOnError((error, stackTrace) {
         FlutterError.dumpErrorToConsole(FlutterErrorDetails(
           exception: error,
@@ -10,7 +10,7 @@ extension StreamFlutterExt<T> on Stream<T> {
       });
 }
 
-extension FutureFlutterExt<T> on Future<T> {
+extension FutureFlutterExtensions<T> on Future<T> {
   Future<T> dumpErrorToConsoleFlutter() => catchError((error, stackTrace) {
         FlutterError.dumpErrorToConsole(FlutterErrorDetails(
           exception: error,