+++ pure_extensions/lib/src/dart/future.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-extension FutureDartExtensions<T> on Future<T> {
+extension FuturePureExtensions<T> on Future<T> {
   /// [Future.delayed]
   Future<T> delay(Duration duration, [FutureOr<T> Function()? computation]) {
     return whenComplete(() => Future.delayed(duration, computation));
@@ -8,4 +8,10 @@ extension FutureDartExtensions<T> on Future<T> {
 
   /// It does not wait for the current future to be concluded
   void unawaited() => this;
+
+  Future<T> onDone(void Function(T value) onDone) async {
+    final value = await this;
+    onDone(value);
+    return value;
+  }
 }