+++ pure_extensions/lib/src/dart/primitives/numbers.dart
@@ -2,7 +2,7 @@ import 'dart:math' as math;
 
 import 'package:rational/rational.dart';
 
-extension NumDartExtensions on num {
+extension NumPureExtensions on num {
   Rational toRational() => Rational.parse(toString());
 
   num min(num other) => math.min<num>(this, other);
@@ -25,7 +25,7 @@ extension NumDartExtensions on num {
   num toDeg() => this * (180.0 / math.pi);
 }
 
-extension DoubleDartExtensions on double {
+extension DoublePureExtensions on double {
   Rational toRational() => Rational.parse(toString());
 
   double min(double other) => math.min<double>(this, other);
@@ -38,7 +38,7 @@ extension DoubleDartExtensions on double {
   double toDeg() => this * (180.0 / math.pi);
 }
 
-extension IntDartExtensions on int {
+extension IntPureExtensions on int {
   Rational toRational() => Rational.fromInt(this);
 
   int min(int other) => math.min<int>(this, other);